2007-07-18 Erik Hjortsberg  <erik.hjortsberg@iteam.se>
	* src/components/ogre/GUICEGUIAdapter.h, src/components/ogre/MotionManager.h, src/services/wfut/WfutSession.h: use size_t where needed

+++ src/components/ogre/GUICEGUIAdapter.h
@@ -80,7 +80,7 @@ private:
 	bool Editbox_selectionChangedHandler(const CEGUI::EventArgs& args);
 	
 	const CEGUI::String* mSelectedText;
-	int mSelectionStart, mSelectionEnd;
+	size_t mSelectionStart, mSelectionEnd;
 };
 
 }
